+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+#### 0.1.9 - December 16, 2015
+* Add additional cursor, implement Window, SMA, StDev via lambdas to ScanLagAllowIncompleteCursor.
+* Make ToSortedMap() live, but unbounded. Could cause out of memory exceptions if used heavily on large series.
+
 #### 0.1.8 - December 11, 2015
 * Optimize cursors further. Virtual methods are 5x faster than Func<>s.
 * Fix SCM clone bug